Move SpectrumLedger from ledger.rs to a Spectrum-specific wallet module
